Fixes and changes from company module review
