+++ frontend/app/components/company-new.js
@@ -1,45 +1,43 @@
 import Component from '@ember/component';
 import { inject as service } from '@ember/service';
+import { Promise } from 'rsvp'
 
 export default Component.extend({
   store: service(),
-  i18n: service(),
-  router: service(),
 
   actions: {
     submit(newCompany) {
       return newCompany.save()
-        .then(function() {
-        })
-        .then(() => this.sendAction('submit', newCompany))
+        .then(() =>
+          Promise.all([
+            ...newCompany
+              .get('locations')
+              .map(location => location.save()),
+            ...newCompany
+              .get('employeeQuantities')
+              .map(employeeQuantity => employeeQuantity.save())
+          ])
+        )
         .then(() => this.get('notify').success('Firma wurde erstellt'))
-        .then(function() {
-          newCompany.get('locations').toArray().forEach(function(location) {
-            newCompany.get('locations').pushObject(location);
-            location.save();
-          });
-          newCompany.get('employeeQuantities').toArray().forEach(function(employeeQuantity) {
-            employeeQuantity.save();
-          });
-        });
+        .then(() => this.sendAction('submit', newCompany))
+      // TODO: What if saving the company or its relations fails?
+      //       For example with an empty company name?
     },
 
     addLocations(company) {
-      let location = this.get('store').createRecord('location');
-      location.set('company', company);
+      this.get('store').createRecord('location', { company });
     },
 
     addEmployeeQuantity(company) {
-      let employeeQuantity = this.get('store').createRecord('employee-quantity');
-      employeeQuantity.set('company', company);
+      this.get('store').createRecord('employee-quantity', { company });
     },
 
     deleteLocation(location) {
-      location.destroyRecord();
+      return location.destroyRecord();
     },
 
     deleteEmployeeQuantity(quantity) {
-      quantity.destroyRecord();
+      return quantity.destroyRecord();
     }
   }
 });
