+++ frontend/app/routes/companies/index.js
@@ -1,0 +1,14 @@
+import Route from '@ember/routing/route';
+
+export default Route.extend({
+  queryParams: {
+    q: {
+      refreshModel: true,
+      replace: true
+    }
+  },
+
+  model({ q }) {
+    return this.store.query('company', { q });
+  },
+});
