+++ frontend/app/routes/companies/new.js
@@ -2,15 +2,12 @@ import Route from '@ember/routing/route';
 
 export default Route.extend({
   model() {
-    let company = this.store.createRecord('company');
-    return company;
+    return this.store.createRecord('company');
   },
 
   actions: {
-    createCompany(company) {
-      let id = company.get('id');
-      this.send('reloadCompaniesList');
-      this.transitionTo('companies.show', id);
+    companyCreated(company) {
+      this.transitionTo('companies.show', company.id);
     }
   }
 });
