+++ frontend/app/templates/companies/new.hbs
@@ -1,2 +1,1 @@
-{{company-new newCompany=model submit=(route-action 'createCompany')}}
-{{outlet}}
+{{company-new newCompany=model submit=(route-action 'companyCreated')}}
