+++ frontend/app/templates/components/company-overview.hbs
