+++ frontend/config/coverage.js
@@ -1,3 +1,6 @@
+/* eslint-env node */
+'use strict';
+
 module.exports = {
   useBabelInstrumenter: true
 }
